feat(share): #MEX-61 implements sharing world WIP
