+++ src/main/resources/public/ts/directives/share-world/share-world.html
@@ -7,6 +7,6 @@
 
     <!-- popup modal to share world -->
     <lightbox class="minetest-lightbox" show="vm.lightbox.share" on-close="vm.closeShareLightbox()">
-        <share-panel app-prefix="'minetest'" resources="vm.world"></share-panel>
+        <share-panel app-prefix="'minetest'" resources="[vm.world]"></share-panel>
     </lightbox>
 </div>
